Crate pallet_liquid_staking_v2

Liquid staking pallet v2

Overview

This pallet manages the NPoS operations for relay chain asset.

Modules

types

weights

Structs

GenesisConfig
Can be used to configure the genesis state of this pallet.

Pallet
The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageExchangeRate

_GeneratedPrefixForStorageMatchingPool

_GeneratedPrefixForStorageStakingPool

_GeneratedPrefixForStorageStakingSettlementRecords

_GeneratedPrefixForStorageUnstakeQueue

Enums

Call
Contains one variant per dispatchable that can be called by an extrinsic.

Error
Custom dispatch errors of this pallet.

Event
The event emitted by this pallet.

Traits

Config
Configuration trait of this pallet.

Type Definitions

ExchangeRate
The exchange rate between relaychain native asset and the voucher.

MatchingPool
Store total stake amount and unstake amount in each era, And will update when stake/unstake occurred.

Module	Deprecated
Type alias to Pallet, to be used by construct_runtime.

StakingPool
Total amount of staked assets in relaycahin.

StakingSettlementRecords
Records reward or slash during each era.

UnstakeQueue
Manage which we should pay off to.