Trait pallet_liquid_staking_v2::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<Self::Event>;
    type Currency: MultiCurrencyExtended<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance, Amount = Amount>;
    type BridgeOrigin: EnsureOrigin<Self::Origin>;
    type StakingCurrency: Get<CurrencyId>;
    type LiquidCurrency: Get<CurrencyId>;
    type PalletId: Get<PalletId>;
    type XcmTransfer: XcmTransfer<Self::AccountId, Balance, CurrencyId>;
    type BaseXcmWeight: Get<Weight>;
    type Agent: Get<MultiLocation>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<Self::Event>

type Currency: MultiCurrencyExtended<Self::AccountId, CurrencyId = CurrencyId, Balance = Balance, Amount = Amount>

Liquid/Staked asset currency.

type BridgeOrigin: EnsureOrigin<Self::Origin>

Offchain bridge accout who manages staking currency in relaychain.

type StakingCurrency: Get<CurrencyId>

The staking currency id.

type LiquidCurrency: Get<CurrencyId>

The liquid voucher currency id.

type PalletId: Get<PalletId>

The pallet id of liquid staking, keeps all the staking assets.

type XcmTransfer: XcmTransfer<Self::AccountId, Balance, CurrencyId>

XCM transfer

type BaseXcmWeight: Get<Weight>

Base xcm transaction weight

type Agent: Get<MultiLocation>

Account manages the staking assets.

type WeightInfo: WeightInfo

Loading content...