Trait pallet_liquid_staking_v2::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Assets: Transfer<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId>;
    type BridgeOrigin: EnsureOrigin<Self::Origin>;
    type UpdateOrigin: EnsureOrigin<Self::Origin>;
    type PalletId: Get<PalletId>;
    type XcmTransfer: XcmTransfer<Self::AccountId, BalanceOf<Self>, AssetIdOf<Self>>;
    type BaseXcmWeight: Get<Weight>;
    type RelayAgent: Get<MultiLocation>;
    type PeriodBasis: Get<BlockNumberFor<Self>>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Assets: Transfer<Self::AccountId> + Inspect<Self::AccountId> + Mutate<Self::AccountId>

Assets for deposit/withdraw assets to/from pallet account

type BridgeOrigin: EnsureOrigin<Self::Origin>

Offchain bridge accout who manages staking currency in relaychain.

type UpdateOrigin: EnsureOrigin<Self::Origin>

The origin which can update liquid currency, staking currency

type PalletId: Get<PalletId>

The pallet id of liquid staking, keeps all the staking assets.

type XcmTransfer: XcmTransfer<Self::AccountId, BalanceOf<Self>, AssetIdOf<Self>>

XCM transfer

type BaseXcmWeight: Get<Weight>

Base xcm transaction weight

type RelayAgent: Get<MultiLocation>

Account manages the staking assets.

type PeriodBasis: Get<BlockNumberFor<Self>>

Basis of period.

type WeightInfo: WeightInfo